requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
sqlparse==0.5.5
tzdata==2025.3
psycopg2-binary==2.9.9

redis==5.0.3 #
django-redis==5.4.0 #

celery
django-celery-beat